requirements.txt
asn1crypto==0.22.0
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==1.9
Django==1.11.7
django-cors-headers==2.1.0
django-filter==1.1.0
django-geojson==2.10.0
django-leaflet==0.22.0
djangorestframework==3.7.1
djangorestframework-gis==0.12
enum34==1.1.6
html5lib==0.999999999
httpie==0.9.9
idna==2.6
ipaddress==1.0.17
itypes==1.1.0
Jinja2==2.9.6
keyring==10.4.0
keyrings.alt==2.2
lxml==4.0.0
Markdown==2.6.9
MarkupSafe==1.0
olefile==0.44
pexpect==4.2.1
Pillow==4.1.1
psycopg2==2.7.3.2
pycrypto==2.6.1
Pygments==2.2.0
pyinotify==0.9.6
pytz==2017.3

